Edit 09-challenge-2 and fix solution Makefile
Add text to motivate challenge 2 in episode 09. This helps explain the
purpose of making an archive file.

Add text to give more detailed instructions on how to solve the
challenge. There are a lot of steps in this challenge!

Fix the `books` directory not being copied into the archive file
correctly in the solution.

Issue: /issues/105
Pull request review: /issues/106